requirements.txt
streamlit
pymysql
ansible-runner
mysql-connector-python
python-dotenv
plotly
pandas
